modified controllers structure
